unsize

Trait CoerciblePtr

pub unsafe trait CoerciblePtr<U: ?Sized>: Sized {
    type Pointee;
    type Output;

    // Required methods
    fn as_sized_ptr(&mut self) -> *mut Self::Pointee;
    unsafe fn replace_ptr(self, _: *mut U) -> Self::Output;
}

Add unsizing methods to pointer-like types.

Safety

A correct implementation must uphold, when calling replace_ptr with valid arguments, that the pointer target and the provenance of the pointer stay unchanged. This allows calling the coercion of inner fields of wrappers even when an invariant depends on the pointer target.

Required Associated Types

type Pointee

The type we point to. This influences which kinds of unsizing are possible.

type Output

The output type when unsizing the pointee to U.

Required Methods

fn as_sized_ptr(&mut self) -> *mut Self::Pointee

Get the raw inner pointer.

unsafe fn replace_ptr(self, _: *mut U) -> Self::Output

Replace the container inner pointer with an unsized version.

Safety

The caller guarantees that the replacement is the same pointer, just a fat pointer variant with a correct tag.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.